@@ -27,17 +27,28 @@ abstract class DbDumper
 
     protected string $dumpBinaryPath = '';
 
+    /** @var array<int, string> */
     protected array $includeTables = [];
 
+    /** @var array<int, string> */
     protected array $excludeTables = [];
 
+    /** @var array<int, string> */
     protected array $extraOptions = [];
 
+    /** @var array<int, string> */
     protected array $extraOptionsAfterDbName = [];
 
     protected bool $appendMode = false;
 
-    protected ?object $compressor = null;
+    protected bool $createTables = true;
+
+    protected bool $includeData = true;
+
+    protected ?Compressor $compressor = null;
+
+    /** @var false|resource */
+    protected mixed $tempFileHandle = false;
 
     public static function create(): static
     {
@@ -49,7 +60,7 @@ abstract class DbDumper
         return $this->dbName;
     }
 
-    public function setDbName(string $dbName): self
+    public function setDbName(string $dbName): static
     {
         $this->dbName = $dbName;
 
@@ -61,7 +72,7 @@ abstract class DbDumper
         return $this->databaseUrl;
     }
 
-    public function setDatabaseUrl(string $databaseUrl): self
+    public function setDatabaseUrl(string $databaseUrl): static
     {
         $this->databaseUrl = $databaseUrl;
 
@@ -70,21 +81,21 @@ abstract class DbDumper
         return $this;
     }
 
-    public function setUserName(string $userName): self
+    public function setUserName(string $userName): static
     {
         $this->userName = $userName;
 
         return $this;
     }
 
-    public function setPassword(string $password): self
+    public function setPassword(string $password): static
     {
         $this->password = $password;
 
         return $this;
     }
 
-    public function setHost(string $host): self
+    public function setHost(string $host): static
     {
         $this->host = $host;
 
@@ -96,28 +107,28 @@ abstract class DbDumper
         return $this->host;
     }
 
-    public function setPort(int $port): self
+    public function setPort(int $port): static
     {
         $this->port = $port;
 
         return $this;
     }
 
-    public function setSocket(string $socket): self
+    public function setSocket(string $socket): static
     {
         $this->socket = $socket;
 
         return $this;
     }
 
-    public function setTimeout(int $timeout): self
+    public function setTimeout(int $timeout): static
     {
         $this->timeout = $timeout;
 
         return $this;
     }
 
-    public function setDumpBinaryPath(string $dumpBinaryPath = ''): self
+    public function setDumpBinaryPath(string $dumpBinaryPath = ''): static
     {
         if ($dumpBinaryPath !== '' && ! str_ends_with($dumpBinaryPath, '/')) {
             $dumpBinaryPath .= '/';
@@ -130,10 +141,10 @@ abstract class DbDumper
 
     public function getCompressorExtension(): string
     {
-        return $this->compressor->useExtension();
+        return $this->compressor?->useExtension() ?? '';
     }
 
-    public function useCompressor(Compressor $compressor): self
+    public function useCompressor(Compressor $compressor): static
     {
         if ($this->appendMode) {
             throw CannotSetParameter::conflictingParameters('compressor', 'append mode');
@@ -144,7 +155,8 @@ abstract class DbDumper
         return $this;
     }
 
-    public function includeTables(string | array $includeTables): self
+    /** @param string|array<int, string> $includeTables */
+    public function includeTables(string|array $includeTables): static
     {
         if (! empty($this->excludeTables)) {
             throw CannotSetParameter::conflictingParameters('includeTables', 'excludeTables');
@@ -159,7 +171,8 @@ abstract class DbDumper
         return $this;
     }
 
-    public function excludeTables(string | array $excludeTables): self
+    /** @param string|array<int, string> $excludeTables */
+    public function excludeTables(string|array $excludeTables): static
     {
         if (! empty($this->includeTables)) {
             throw CannotSetParameter::conflictingParameters('excludeTables', 'includeTables');
@@ -174,27 +187,41 @@ abstract class DbDumper
         return $this;
     }
 
-    public function addExtraOption(string $extraOption): self
+    public function addExtraOption(string $extraOption): static
     {
-        if (! empty($extraOption)) {
+        if ($extraOption !== '') {
             $this->extraOptions[] = $extraOption;
         }
 
         return $this;
     }
 
-    public function addExtraOptionAfterDbName(string $extraOptionAfterDbName): self
+    public function addExtraOptionAfterDbName(string $extraOptionAfterDbName): static
     {
-        if (! empty($extraOptionAfterDbName)) {
+        if ($extraOptionAfterDbName !== '') {
             $this->extraOptionsAfterDbName[] = $extraOptionAfterDbName;
         }
 
         return $this;
     }
 
+    public function doNotCreateTables(): static
+    {
+        $this->createTables = false;
+
+        return $this;
+    }
+
+    public function doNotDumpData(): static
+    {
+        $this->includeData = false;
+
+        return $this;
+    }
+
     abstract public function dumpToFile(string $dumpFile): void;
 
-    public function checkIfDumpWasSuccessFul(Process $process, string $outputFile): void
+    public function checkIfDumpWasSuccessful(Process $process, string $outputFile): void
     {
         if (! $process->isSuccessful()) {
             throw DumpFailed::processDidNotEndSuccessfully($process);
@@ -207,6 +234,18 @@ abstract class DbDumper
         if (filesize($outputFile) === 0) {
             throw DumpFailed::dumpfileWasEmpty($process);
         }
+    }
+
+    /** @return false|resource */
+    public function getTempFileHandle(): mixed
+    {
+        return $this->tempFileHandle;
+    }
+
+    /** @param false|resource $tempFileHandle */
+    public function setTempFileHandle(mixed $tempFileHandle): void
+    {
+        $this->tempFileHandle = $tempFileHandle;
     }
 
     protected function configureFromDatabaseUrl(): void
@@ -229,14 +268,19 @@ abstract class DbDumper
                     continue;
                 }
 
-                $this->$setterMethod($value);
+                if ($setterMethod === 'setPort') {
+                    $this->setPort(is_numeric($value) ? (int) $value : 0);
+                } else {
+                    $stringValue = is_scalar($value) ? (string) $value : '';
+                    $this->$setterMethod($stringValue);
+                }
             }
         }
     }
 
     protected function getCompressCommand(string $command, string $dumpFile): string
     {
-        $compressCommand = $this->compressor->useCommand();
+        $compressCommand = $this->compressor?->useCommand() ?? '';
 
         if ($this->isWindows()) {
             return "{$command} | {$compressCommand} > {$dumpFile}";
@@ -245,7 +289,7 @@ abstract class DbDumper
         return "(((({$command}; echo \$? >&3) | {$compressCommand} > {$dumpFile}) 3>&1) | (read x; exit \$x))";
     }
 
-    protected function echoToFile(string $command, string $dumpFile): string
+    protected function redirectCommandOutput(string $command, string $dumpFile): string
     {
         $dumpFile = '"' . addcslashes($dumpFile, '\\"') . '"';
 
