@@ -6,8 +6,8 @@ use Exception;
 
 class CannotSetParameter extends Exception
 {
-    public static function conflictingParameters(string $name, string $conflictName): static
+    public static function conflictingParameters(string $name, string $conflictName): self
     {
-        return new static("Cannot set `{$name}` because it conflicts with parameter `{$conflictName}`.");
+        return new self("Cannot set `{$name}` because it conflicts with parameter `{$conflictName}`.");
     }
 }
