@@ -6,8 +6,8 @@ use Exception;
 
 class CannotStartDump extends Exception
 {
-    public static function emptyParameter(string $name): static
+    public static function emptyParameter(string $name): self
     {
-        return new static("Parameter `{$name}` cannot be empty.");
+        return new self("Parameter `{$name}` cannot be empty.");
     }
 }
