@@ -7,25 +7,25 @@ use Symfony\Component\Process\Process;
 
 class DumpFailed extends Exception
 {
-    public static function processDidNotEndSuccessfully(Process $process): static
+    public static function processDidNotEndSuccessfully(Process $process): self
     {
         $processOutput = static::formatProcessOutput($process);
 
-        return new static("The dump process failed with a none successful exitcode.{$processOutput}");
+        return new self("The dump process failed with a none successful exitcode.{$processOutput}");
     }
 
-    public static function dumpfileWasNotCreated(Process $process): static
+    public static function dumpfileWasNotCreated(Process $process): self
     {
         $processOutput = static::formatProcessOutput($process);
 
-        return new static("The dumpfile could not be created.{$processOutput}");
+        return new self("The dumpfile could not be created.{$processOutput}");
     }
 
-    public static function dumpfileWasEmpty(Process $process): static
+    public static function dumpfileWasEmpty(Process $process): self
     {
         $processOutput = static::formatProcessOutput($process);
 
-        return new static("The created dumpfile is empty.{$processOutput}");
+        return new self("The created dumpfile is empty.{$processOutput}");
     }
 
     protected static function formatProcessOutput(Process $process): string
