@@ -10,9 +10,9 @@ it('provides a factory method')
     ->expect(MariaDb::create())
     ->toBeInstanceOf(MariaDb::class);
 
-it('will throw an exception when no credentials are set')
-    ->tap(fn () => MariaDb::create()->dumpToFile('test.sql'))
-    ->throws(CannotStartDump::class);
+it('will throw an exception when no credentials are set', function () {
+    MariaDb::create()->dumpToFile('test.sql');
+})->throws(CannotStartDump::class);
 
 it('can generate a dump command', function () {
     $dumpCommand = MariaDb::create()
@@ -105,7 +105,7 @@ it('can generate a dump command without using comments', function () {
         ->setDbName('dbname')
         ->setUserName('username')
         ->setPassword('password')
-        ->dontSkipComments()
+        ->doNotSkipComments()
         ->getDumpCommand('dump.sql', 'credentials.txt');
 
     expect($dumpCommand)->toEqual(
@@ -118,7 +118,7 @@ it('can generate a dump command without using extended inserts', function () {
         ->setDbName('dbname')
         ->setUserName('username')
         ->setPassword('password')
-        ->dontUseExtendedInserts()
+        ->doNotUseExtendedInserts()
         ->getDumpCommand('dump.sql', 'credentials.txt');
 
     expect($dumpCommand)->toEqual(
@@ -144,7 +144,7 @@ it('can generate a dump command without using extending inserts', function () {
         ->setDbName('dbname')
         ->setUserName('username')
         ->setPassword('password')
-        ->dontUseExtendedInserts()
+        ->doNotUseExtendedInserts()
         ->getDumpCommand('dump.sql', 'credentials.txt');
 
     expect($dumpCommand)->toEqual(
@@ -233,7 +233,7 @@ it('can generate a dump command not skipping auto increment values', function ()
         ->setDbName('dbname')
         ->setUserName('username')
         ->setPassword('password')
-        ->dontSkipAutoIncrement()
+        ->doNotSkipAutoIncrement()
         ->getDumpCommand('dump.sql', 'credentials.txt');
 
     expect($dumpCommand)->not->toContain("sed 's/ AUTO_INCREMENT=[0-9]*\\b//'");
@@ -473,5 +473,32 @@ it('can generate a dump command with no data', function () {
 
     expect($dumpCommand)->toEqual(
         '\'mariadb-dump\' --defaults-extra-file="credentials.txt" --no-data --skip-comments --extended-insert dbname > "dump.sql"'
+    );
+});
+
+it('can generate a dump command with routines included', function () {
+    $dumpCommand = MariaDb::create()
+        ->setDbName('dbname')
+        ->setUserName('username')
+        ->setPassword('password')
+        ->includeRoutines()
+        ->getDumpCommand('dump.sql', 'credentials.txt');
+
+    expect($dumpCommand)->toEqual(
+        '\'mariadb-dump\' --defaults-extra-file="credentials.txt" --skip-comments --extended-insert --routines dbname > "dump.sql"'
+    );
+});
+
+it('can generate a dump command excluding data for specific tables', function () {
+    $dumpCommand = MariaDb::create()
+        ->setDbName('dbname')
+        ->setUserName('username')
+        ->setPassword('password')
+        ->excludeTablesData(['tb1', 'tb2'])
+        ->getDumpCommand('dump.sql', 'credentials.txt');
+
+    expect($dumpCommand)->toEqual(
+        '\'mariadb-dump\' --defaults-extra-file="credentials.txt" --skip-comments --extended-insert ' .
+        '--ignore-table-data=dbname.tb1 --ignore-table-data=dbname.tb2 dbname > "dump.sql"'
     );
 });
